- Upgrade project to Android API 33 - Added Android X support - Update project to Gradle 7.4 - Update to use maven publishing - Move HttpUtils out of library to avoid Unsafe SSL TrustManager and Unsafe HostnameVerifier Google Play warnings - Upgrade to OkHttp 4.10.0
