+++ okhttputils/build.gradle
@@ -1,16 +1,18 @@
-apply plugin: 'com.android.library'
-apply plugin: 'com.novoda.bintray-release'//添加
+plugins {
+    id 'com.android.library'
+}
 
+apply from: "$rootProject.projectDir/publishing.gradle"
 
 android {
-    compileSdkVersion 23
-    buildToolsVersion "23.0.1"
+    compileSdkVersion 33
+    buildToolsVersion "33.0.0"
 
     defaultConfig {
-        minSdkVersion 9
-        targetSdkVersion 23
+        minSdkVersion 14
+        targetSdkVersion 33
         versionCode 1
-        versionName "1.0"
+        versionName "2.6.3"
     }
     buildTypes {
         release {
@@ -23,8 +25,10 @@ android {
         warning 'InvalidPackage'
     }
 
+    publishing {
+        singleVariant('release')
+    }
 }
-
 
 task clearJar(type: Delete) {
     delete 'build/outputs/okhttputils.jar'
@@ -39,20 +43,46 @@ task makeJar(type: Copy) {
 
 makeJar.dependsOn(clearJar, build)
 
+afterEvaluate {
+    publishing {
+        publications {
+            release(MavenPublication) {
+                from components.release
 
+                groupId = 'com.zhy'
+                artifactId = 'okhttputils'
+                version = "$project.android.defaultConfig.versionName"
+
+                pom {
+                    name = 'okhttputils'
+                    description = 'Oh hi, this is a nice description for a project, right?'
+                    url = 'https://github.com/hongyangAndroid/okhttputils'
+                    packaging = 'aar'
+                    licenses {
+                        license {
+                            name = 'Apache License, Version 2.0'
+                            url = 'http://www.apache.org/licenses/LICENSE-2.0'
+                            distribution = 'repo'
+                        }
+                    }
+                    developers {
+                        developer {
+                            id = 'hongyangandroid'
+                            name = 'Hongyang Android'
+                        }
+                    }
+                    scm {
+                        connection = 'scm:hg:https://github.com/hongyangAndroid/okhttputils'
+                        developerConnection = 'scm:hg:https://github.com/hongyangAndroid/okhttputils'
+                        url = 'https://github.com/hongyangAndroid/okhttputils'
+                    }
+                }
+            }
+        }
+    }
+}
 
 dependencies {
-    compile fileTree(dir: 'libs', include: ['*.jar'])
-    compile 'com.squareup.okhttp3:okhttp:3.3.1'
+    implementation fileTree(dir: 'libs', include: ['*.jar'])
+    implementation 'com.squareup.okhttp3:okhttp:4.10.0'
 }
-
-//添加
-publish {
-    userOrg = 'hongyangandroid'//bintray.com用户名
-    groupId = 'com.zhy'//jcenter上的路径
-    artifactId = 'okhttputils'//项目名称
-    publishVersion = '2.6.2'//版本号
-    desc = 'Oh hi, this is a nice description for a project, right?'
-    website = 'https://github.com/hongyangAndroid/okhttp-utils'
-}
-
