+++ sample-okhttp/build.gradle
@@ -1,13 +1,15 @@
-apply plugin: 'com.android.application'
+plugins {
+    id 'com.android.application'
+}
 
 android {
-    compileSdkVersion 22
-    buildToolsVersion "22.0.1"
+    compileSdkVersion 33
+    buildToolsVersion "33.0.0"
 
     defaultConfig {
         applicationId "com.zhy.sample_okhttp"
-        minSdkVersion 10
-        targetSdkVersion 22
+        minSdkVersion 14
+        targetSdkVersion 33
         versionCode 1
         versionName "1.0"
     }
@@ -25,11 +27,11 @@ android {
 }
 
 dependencies {
-    compile fileTree(dir: 'libs', include: ['*.jar'])
-    compile 'com.android.support:appcompat-v7:22.2.1'
-    compile project(':okhttputils')
-//    compile 'com.zhy:okhttputils:2.6.1'
-    compile 'com.google.code.gson:gson:2.3.1'
-    compile 'com.squareup.okhttp3:okhttp-urlconnection:3.2.0'
-    compile 'com.github.franmontiel:PersistentCookieJar:v0.9.3'
+    implementation fileTree(dir: 'libs', include: ['*.jar'])
+    implementation project(':okhttputils')
+
+    implementation 'androidx.appcompat:appcompat:1.5.1'
+    implementation 'com.google.code.gson:gson:2.9.0'
+    implementation 'com.squareup.okhttp3:okhttp-urlconnection:3.2.0'
+    implementation 'com.github.franmontiel:PersistentCookieJar:v1.0.1'
 }
