+++ sample-okhttp/src/main/java/com/zhy/sample_okhttp/MainActivity.java
@@ -3,13 +3,14 @@ package com.zhy.sample_okhttp;
 import android.graphics.Bitmap;
 import android.os.Bundle;
 import android.os.Environment;
-import android.support.v7.app.AppCompatActivity;
 import android.util.Log;
 import android.view.View;
 import android.widget.ImageView;
 import android.widget.ProgressBar;
 import android.widget.TextView;
 import android.widget.Toast;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.google.gson.Gson;
 import com.zhy.http.okhttp.OkHttpUtils;
