+++ settings.gradle
@@ -1,1 +1,18 @@
-include ':okhttputils', ':sample-okhttp'
+pluginManagement {
+    repositories {
+        gradlePluginPortal()
+        google()
+        mavenCentral()
+    }
+}
+/*dependencyResolutionManagement {
+    repositoriesMode.set(RepositoriesMode.FAIL_ON_PROJECT_REPOS)
+    repositories {
+        google()
+        mavenCentral()
+    }
+}*/
+
+
+include ':okhttputils'
+include ':sample-okhttp'
